Move GetNodeIP() to netutils
