+++ pkg/netutils/common.go
@@ -2,6 +2,7 @@ package netutils
 
 import (
 	"encoding/binary"
+	"fmt"
 	"net"
 )
 
@@ -19,4 +20,24 @@ func Uint32ToIP(u uint32) net.IP {
 func GenerateDefaultGateway(sna *net.IPNet) net.IP {
 	ip := sna.IP.To4()
 	return net.IPv4(ip[0], ip[1], ip[2], ip[3]|0x1)
+}
+
+func GetNodeIP(nodeName string) (string, error) {
+	ip := net.ParseIP(nodeName)
+	if ip == nil {
+		addrs, err := net.LookupIP(nodeName)
+		if err != nil {
+			return "", fmt.Errorf("Failed to lookup IP address for node %s: %v", nodeName, err)
+		}
+		for _, addr := range addrs {
+			if addr.String() != "127.0.0.1" {
+				ip = addr
+				break
+			}
+		}
+	}
+	if ip == nil || len(ip.String()) == 0 {
+		return "", fmt.Errorf("Failed to obtain IP address from node name: %s", nodeName)
+	}
+	return ip.String(), nil
 }
