+++ plugins/osdn/osdn.go
@@ -21,7 +21,7 @@ import (
 	"k8s.io/kubernetes/pkg/types"
 	"k8s.io/kubernetes/pkg/watch"
 
-	osdn "github.com/openshift/openshift-sdn/pkg/ovssubnet"
+	"github.com/openshift/openshift-sdn/pkg/netutils"
 	osdnapi "github.com/openshift/openshift-sdn/pkg/ovssubnet/api"
 
 	osclient "github.com/openshift/origin/pkg/client"
@@ -193,7 +193,7 @@ func (oi *OsdnRegistryInterface) GetNodes() ([]osdnapi.Node, string, error) {
 			nodeIP = node.Status.Addresses[0].Address
 		} else {
 			var err error
-			nodeIP, err = osdn.GetNodeIP(node.ObjectMeta.Name)
+			nodeIP, err = netutils.GetNodeIP(node.ObjectMeta.Name)
 			if err != nil {
 				return nil, "", err
 			}
@@ -238,7 +238,7 @@ func (oi *OsdnRegistryInterface) WatchNodes(receiver chan<- *osdnapi.NodeEvent, 
 		if len(node.Status.Addresses) > 0 {
 			nodeIP = node.Status.Addresses[0].Address
 		} else {
-			nodeIP, err = osdn.GetNodeIP(node.ObjectMeta.Name)
+			nodeIP, err = netutils.GetNodeIP(node.ObjectMeta.Name)
 			if err != nil {
 				return err
 			}
